refactor: move code to separate files
Refactor and move code to separate files to improve readability.

This refactor also moves batch to a separate file, but as most logic
is defined on *conn, this logic is not yet moved. This will be done
in a follow-up pull request.
