+++ driver.go
@@ -21,19 +21,14 @@ import (
 	"fmt"
 	"io"
 	"log/slog"
-	"math/big"
-	"reflect"
 	"regexp"
-	"slices"
 	"strconv"
 	"strings"
 	"sync"
 	"time"
 
-	"cloud.google.com/go/civil"
 	"cloud.google.com/go/spanner"
 	adminapi "cloud.google.com/go/spanner/admin/database/apiv1"
-	adminpb "cloud.google.com/go/spanner/admin/database/apiv1/databasepb"
 	"cloud.google.com/go/spanner/apiv1/spannerpb"
 	"github.com/google/uuid"
 	"github.com/googleapis/gax-go/v2"
@@ -43,8 +38,6 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/grpc/status"
-	"google.golang.org/protobuf/proto"
-	"google.golang.org/protobuf/reflect/protoreflect"
 )
 
 const userAgent = "go-sql-spanner/1.9.0" // x-release-please-version
@@ -91,36 +84,6 @@ var _ driver.DriverContext = &Driver{}
 func init() {
 	sql.Register("spanner", &Driver{connectors: make(map[string]*connector)})
 }
-
-// ExecOptions can be passed in as an argument to the Query, QueryContext,
-// Exec, and ExecContext functions to specify additional execution options
-// for a statement.
-type ExecOptions struct {
-	// DecodeOption indicates how the returned rows should be decoded.
-	DecodeOption DecodeOption
-
-	// TransactionOptions are the transaction options that will be used for
-	// the transaction that is started by the statement.
-	TransactionOptions spanner.TransactionOptions
-	// QueryOptions are the query options that will be used for the statement.
-	QueryOptions spanner.QueryOptions
-}
-
-type DecodeOption int
-
-const (
-	// DecodeOptionNormal decodes into idiomatic Go types (e.g. bool, string, int64, etc.)
-	DecodeOptionNormal DecodeOption = iota
-
-	// DecodeOptionProto does not decode the returned rows at all, and instead just returns
-	// the underlying protobuf objects. Use this for advanced use-cases where you want
-	// direct access to the underlying values.
-	// All values should be scanned into an instance of spanner.GenericColumnValue like this:
-	//
-	// 	var v spanner.GenericColumnValue
-	// 	row.Scan(&v)
-	DecodeOptionProto
-)
 
 // Driver represents a Google Cloud Spanner database/sql driver.
 type Driver struct {
@@ -630,921 +593,6 @@ func runTransactionWithOptions(ctx context.Context, db *sql.DB, opts *sql.TxOpti
 	}
 }
 
-// SpannerConn is the public interface for the raw Spanner connection for the
-// sql driver. This interface can be used with the db.Conn().Raw() method.
-type SpannerConn interface {
-	// StartBatchDDL starts a DDL batch on the connection. After calling this
-	// method all subsequent DDL statements will be cached locally. Calling
-	// RunBatch will send all cached DDL statements to Spanner as one batch.
-	// Use DDL batching to speed up the execution of multiple DDL statements.
-	// Note that a DDL batch is not atomic. It is possible that some DDL
-	// statements are executed successfully and some not.
-	// See https://cloud.google.com/spanner/docs/schema-updates#order_of_execution_of_statements_in_batches
-	// for more information on how Cloud Spanner handles DDL batches.
-	StartBatchDDL() error
-	// StartBatchDML starts a DML batch on the connection. After calling this
-	// method all subsequent DML statements will be cached locally. Calling
-	// RunBatch will send all cached DML statements to Spanner as one batch.
-	// Use DML batching to speed up the execution of multiple DML statements.
-	// DML batches can be executed both outside of a transaction and during
-	// a read/write transaction. If a DML batch is executed outside an active
-	// transaction, the batch will be applied atomically to the database if
-	// successful and rolled back if one or more of the statements fail.
-	// If a DML batch is executed as part of a transaction, the error will
-	// be returned to the application, and the application can decide whether
-	// to commit or rollback the transaction.
-	StartBatchDML() error
-	// RunBatch sends all batched DDL or DML statements to Spanner. This is a
-	// no-op if no statements have been batched or if there is no active batch.
-	RunBatch(ctx context.Context) error
-	// AbortBatch aborts the current DDL or DML batch and discards all batched
-	// statements.
-	AbortBatch() error
-	// InDDLBatch returns true if the connection is currently in a DDL batch.
-	InDDLBatch() bool
-	// InDMLBatch returns true if the connection is currently in a DML batch.
-	InDMLBatch() bool
-	// GetBatchedStatements returns a copy of the statements that are currently
-	// buffered to be executed as a DML or DDL batch. It returns an empty slice
-	// if no batch is active, or if there are no statements buffered.
-	GetBatchedStatements() []spanner.Statement
-
-	// RetryAbortsInternally returns true if the connection automatically
-	// retries all aborted transactions.
-	RetryAbortsInternally() bool
-	// SetRetryAbortsInternally enables/disables the automatic retry of aborted
-	// transactions. If disabled, any aborted error from a transaction will be
-	// propagated to the application.
-	SetRetryAbortsInternally(retry bool) error
-
-	// AutocommitDMLMode returns the current mode that is used for DML
-	// statements outside a transaction. The default is Transactional.
-	AutocommitDMLMode() AutocommitDMLMode
-	// SetAutocommitDMLMode sets the mode to use for DML statements that are
-	// executed outside transactions. The default is Transactional. Change to
-	// PartitionedNonAtomic to use Partitioned DML instead of Transactional DML.
-	// See https://cloud.google.com/spanner/docs/dml-partitioned for more
-	// information on Partitioned DML.
-	SetAutocommitDMLMode(mode AutocommitDMLMode) error
-
-	// ReadOnlyStaleness returns the current staleness that is used for
-	// queries in autocommit mode, and for read-only transactions.
-	ReadOnlyStaleness() spanner.TimestampBound
-	// SetReadOnlyStaleness sets the staleness to use for queries in autocommit
-	// mode and for read-only transaction.
-	SetReadOnlyStaleness(staleness spanner.TimestampBound) error
-
-	// TransactionTag returns the transaction tag that will be applied to the next
-	// read/write transaction on this connection. The transaction tag that is set
-	// on the connection is cleared when a read/write transaction is started.
-	TransactionTag() string
-	// SetTransactionTag sets the transaction tag that should be applied to the
-	// next read/write transaction on this connection. The tag is cleared when a
-	// read/write transaction is started.
-	SetTransactionTag(transactionTag string) error
-
-	// ExcludeTxnFromChangeStreams returns true if the next transaction should be excluded from change streams with the
-	// DDL option `allow_txn_exclusion=true`.
-	ExcludeTxnFromChangeStreams() bool
-	// SetExcludeTxnFromChangeStreams sets whether the next transaction should be excluded from change streams with the
-	// DDL option `allow_txn_exclusion=true`.
-	SetExcludeTxnFromChangeStreams(excludeTxnFromChangeStreams bool) error
-
-	// Apply writes an array of mutations to the database. This method may only be called while the connection
-	// is outside a transaction. Use BufferWrite to write mutations in a transaction.
-	// See also spanner.Client#Apply
-	Apply(ctx context.Context, ms []*spanner.Mutation, opts ...spanner.ApplyOption) (commitTimestamp time.Time, err error)
-
-	// BufferWrite writes an array of mutations to the current transaction. This method may only be called while the
-	// connection is in a read/write transaction. Use Apply to write mutations outside a transaction.
-	// See also spanner.ReadWriteTransaction#BufferWrite
-	BufferWrite(ms []*spanner.Mutation) error
-
-	// CommitTimestamp returns the commit timestamp of the last implicit or explicit read/write transaction that
-	// was executed on the connection, or an error if the connection has not executed a read/write transaction
-	// that committed successfully. The timestamp is in the local timezone.
-	CommitTimestamp() (commitTimestamp time.Time, err error)
-
-	// UnderlyingClient returns the underlying Spanner client for the database.
-	// The client cannot be used to access the current transaction or batch on
-	// this connection. Executing a transaction or batch using the client that is
-	// returned, does not affect this connection.
-	// Note that multiple connections share the same Spanner client. Calling
-	// this function on different connections to the same database, can
-	// return the same Spanner client.
-	UnderlyingClient() (client *spanner.Client, err error)
-
-	// resetTransactionForRetry resets the current transaction after it has
-	// been aborted by Spanner. Calling this function on a transaction that
-	// has not been aborted is not supported and will cause an error to be
-	// returned.
-	resetTransactionForRetry(ctx context.Context, errDuringCommit bool) error
-
-	// setTransactionOptions sets the TransactionOptions that should be used
-	// for this transaction.
-	withTransactionOptions(options spanner.TransactionOptions)
-}
-
-var _ SpannerConn = &conn{}
-
-type conn struct {
-	connector     *connector
-	closed        bool
-	client        *spanner.Client
-	adminClient   *adminapi.DatabaseAdminClient
-	connId        string
-	logger        *slog.Logger
-	tx            contextTransaction
-	prevTx        contextTransaction
-	resetForRetry bool
-	commitTs      *time.Time
-	database      string
-	retryAborts   bool
-
-	execSingleQuery              func(ctx context.Context, c *spanner.Client, statement spanner.Statement, bound spanner.TimestampBound, options ExecOptions) *spanner.RowIterator
-	execSingleQueryTransactional func(ctx context.Context, c *spanner.Client, statement spanner.Statement, options ExecOptions) (rowIterator, time.Time, error)
-	execSingleDMLTransactional   func(ctx context.Context, c *spanner.Client, statement spanner.Statement, options ExecOptions) (int64, time.Time, error)
-	execSingleDMLPartitioned     func(ctx context.Context, c *spanner.Client, statement spanner.Statement, options ExecOptions) (int64, error)
-
-	// batch is the currently active DDL or DML batch on this connection.
-	batch *batch
-
-	// autocommitDMLMode determines the type of DML to use when a single DML
-	// statement is executed on a connection. The default is Transactional, but
-	// it can also be set to PartitionedNonAtomic to execute the statement as
-	// Partitioned DML.
-	autocommitDMLMode AutocommitDMLMode
-	// readOnlyStaleness is used for queries in autocommit mode and for read-only transactions.
-	readOnlyStaleness spanner.TimestampBound
-
-	// execOptions are applied to the next statement that is executed on this connection.
-	// It can be set by passing it in as an argument to ExecContext or QueryContext
-	// and is cleared after each execution.
-	execOptions ExecOptions
-}
-
-type batchType int
-
-const (
-	ddl batchType = iota
-	dml
-)
-
-type batch struct {
-	tp         batchType
-	statements []spanner.Statement
-	options    ExecOptions
-}
-
-// AutocommitDMLMode indicates whether a single DML statement should be executed
-// in a normal atomic transaction or as a Partitioned DML statement.
-// See https://cloud.google.com/spanner/docs/dml-partitioned for more information.
-type AutocommitDMLMode int
-
-func (mode AutocommitDMLMode) String() string {
-	switch mode {
-	case Transactional:
-		return "Transactional"
-	case PartitionedNonAtomic:
-		return "Partitioned_Non_Atomic"
-	}
-	return ""
-}
-
-const (
-	Transactional AutocommitDMLMode = iota
-	PartitionedNonAtomic
-)
-
-func (c *conn) UnderlyingClient() (*spanner.Client, error) {
-	return c.client, nil
-}
-
-func (c *conn) CommitTimestamp() (time.Time, error) {
-	if c.commitTs == nil {
-		return time.Time{}, spanner.ToSpannerError(status.Error(codes.FailedPrecondition, "this connection has not executed a read/write transaction that committed successfully"))
-	}
-	return *c.commitTs, nil
-}
-
-func (c *conn) RetryAbortsInternally() bool {
-	return c.retryAborts
-}
-
-func (c *conn) SetRetryAbortsInternally(retry bool) error {
-	_, err := c.setRetryAbortsInternally(retry)
-	return err
-}
-
-func (c *conn) setRetryAbortsInternally(retry bool) (driver.Result, error) {
-	if c.inTransaction() {
-		return nil, spanner.ToSpannerError(status.Error(codes.FailedPrecondition, "cannot change retry mode while a transaction is active"))
-	}
-	c.retryAborts = retry
-	return driver.ResultNoRows, nil
-}
-
-func (c *conn) AutocommitDMLMode() AutocommitDMLMode {
-	return c.autocommitDMLMode
-}
-
-func (c *conn) SetAutocommitDMLMode(mode AutocommitDMLMode) error {
-	_, err := c.setAutocommitDMLMode(mode)
-	return err
-}
-
-func (c *conn) setAutocommitDMLMode(mode AutocommitDMLMode) (driver.Result, error) {
-	c.autocommitDMLMode = mode
-	return driver.ResultNoRows, nil
-}
-
-func (c *conn) ReadOnlyStaleness() spanner.TimestampBound {
-	return c.readOnlyStaleness
-}
-
-func (c *conn) SetReadOnlyStaleness(staleness spanner.TimestampBound) error {
-	_, err := c.setReadOnlyStaleness(staleness)
-	return err
-}
-
-func (c *conn) setReadOnlyStaleness(staleness spanner.TimestampBound) (driver.Result, error) {
-	c.readOnlyStaleness = staleness
-	return driver.ResultNoRows, nil
-}
-
-func (c *conn) ExcludeTxnFromChangeStreams() bool {
-	return c.execOptions.TransactionOptions.ExcludeTxnFromChangeStreams
-}
-
-func (c *conn) SetExcludeTxnFromChangeStreams(excludeTxnFromChangeStreams bool) error {
-	_, err := c.setExcludeTxnFromChangeStreams(excludeTxnFromChangeStreams)
-	return err
-}
-
-func (c *conn) setExcludeTxnFromChangeStreams(excludeTxnFromChangeStreams bool) (driver.Result, error) {
-	if c.inTransaction() {
-		return nil, spanner.ToSpannerError(status.Error(codes.FailedPrecondition, "cannot set ExcludeTxnFromChangeStreams while a transaction is active"))
-	}
-	c.execOptions.TransactionOptions.ExcludeTxnFromChangeStreams = excludeTxnFromChangeStreams
-	return driver.ResultNoRows, nil
-}
-
-func (c *conn) TransactionTag() string {
-	return c.execOptions.TransactionOptions.TransactionTag
-}
-
-func (c *conn) SetTransactionTag(transactionTag string) error {
-	_, err := c.setTransactionTag(transactionTag)
-	return err
-}
-
-func (c *conn) setTransactionTag(transactionTag string) (driver.Result, error) {
-	if c.inTransaction() {
-		return nil, spanner.ToSpannerError(status.Error(codes.FailedPrecondition, "cannot set transaction tag while a transaction is active"))
-	}
-	c.execOptions.TransactionOptions.TransactionTag = transactionTag
-	return driver.ResultNoRows, nil
-}
-
-func (c *conn) StatementTag() string {
-	return c.execOptions.QueryOptions.RequestTag
-}
-
-func (c *conn) SetStatementTag(statementTag string) error {
-	_, err := c.setStatementTag(statementTag)
-	return err
-}
-
-func (c *conn) setStatementTag(statementTag string) (driver.Result, error) {
-	c.execOptions.QueryOptions.RequestTag = statementTag
-	return driver.ResultNoRows, nil
-}
-
-func (c *conn) StartBatchDDL() error {
-	_, err := c.startBatchDDL()
-	return err
-}
-
-func (c *conn) StartBatchDML() error {
-	_, err := c.startBatchDML()
-	return err
-}
-
-func (c *conn) RunBatch(ctx context.Context) error {
-	_, err := c.runBatch(ctx)
-	return err
-}
-
-func (c *conn) AbortBatch() error {
-	_, err := c.abortBatch()
-	return err
-}
-
-func (c *conn) InDDLBatch() bool {
-	return c.batch != nil && c.batch.tp == ddl
-}
-
-func (c *conn) InDMLBatch() bool {
-	return (c.batch != nil && c.batch.tp == dml) || (c.inReadWriteTransaction() && c.tx.(*readWriteTransaction).batch != nil)
-}
-
-func (c *conn) GetBatchedStatements() []spanner.Statement {
-	if c.batch == nil || c.batch.statements == nil {
-		return []spanner.Statement{}
-	}
-	return slices.Clone(c.batch.statements)
-}
-
-func (c *conn) inBatch() bool {
-	return c.InDDLBatch() || c.InDMLBatch()
-}
-
-func (c *conn) startBatchDDL() (driver.Result, error) {
-	if c.batch != nil {
-		return nil, spanner.ToSpannerError(status.Errorf(codes.FailedPrecondition, "This connection already has an active batch."))
-	}
-	if c.inTransaction() {
-		return nil, spanner.ToSpannerError(status.Errorf(codes.FailedPrecondition, "This connection has an active transaction. DDL batches in transactions are not supported."))
-	}
-	c.logger.Debug("started ddl batch")
-	c.batch = &batch{tp: ddl}
-	return driver.ResultNoRows, nil
-}
-
-func (c *conn) startBatchDML() (driver.Result, error) {
-	execOptions := c.options()
-
-	if c.inTransaction() {
-		return c.tx.StartBatchDML(execOptions.QueryOptions)
-	}
-
-	if c.batch != nil {
-		return nil, spanner.ToSpannerError(status.Errorf(codes.FailedPrecondition, "This connection already has an active batch."))
-	}
-	if c.inReadOnlyTransaction() {
-		return nil, spanner.ToSpannerError(status.Errorf(codes.FailedPrecondition, "This connection has an active read-only transaction. Read-only transactions cannot execute DML batches."))
-	}
-	c.logger.Debug("starting dml batch")
-	c.batch = &batch{tp: dml, options: execOptions}
-	return driver.ResultNoRows, nil
-}
-
-func (c *conn) runBatch(ctx context.Context) (driver.Result, error) {
-	if c.inTransaction() {
-		return c.tx.RunBatch(ctx)
-	}
-
-	if c.batch == nil {
-		return nil, spanner.ToSpannerError(status.Errorf(codes.FailedPrecondition, "This connection does not have an active batch"))
-	}
-	switch c.batch.tp {
-	case ddl:
-		return c.runDDLBatch(ctx)
-	case dml:
-		return c.runDMLBatch(ctx)
-	default:
-		return nil, spanner.ToSpannerError(status.Errorf(codes.InvalidArgument, "Unknown batch type: %d", c.batch.tp))
-	}
-}
-
-func (c *conn) runDDLBatch(ctx context.Context) (driver.Result, error) {
-	statements := c.batch.statements
-	c.batch = nil
-	return c.execDDL(ctx, statements...)
-}
-
-func (c *conn) runDMLBatch(ctx context.Context) (driver.Result, error) {
-	statements := c.batch.statements
-	options := c.batch.options
-	c.batch = nil
-	return c.execBatchDML(ctx, statements, options)
-}
-
-func (c *conn) abortBatch() (driver.Result, error) {
-	if c.inTransaction() {
-		return c.tx.AbortBatch()
-	}
-
-	c.batch = nil
-	return driver.ResultNoRows, nil
-}
-
-func (c *conn) execDDL(ctx context.Context, statements ...spanner.Statement) (driver.Result, error) {
-	if c.batch != nil && c.batch.tp == dml {
-		return nil, spanner.ToSpannerError(status.Error(codes.FailedPrecondition, "This connection has an active DML batch"))
-	}
-	if c.batch != nil && c.batch.tp == ddl {
-		c.batch.statements = append(c.batch.statements, statements...)
-		return driver.ResultNoRows, nil
-	}
-
-	if len(statements) > 0 {
-		ddlStatements := make([]string, len(statements))
-		for i, s := range statements {
-			ddlStatements[i] = s.SQL
-		}
-		op, err := c.adminClient.UpdateDatabaseDdl(ctx, &adminpb.UpdateDatabaseDdlRequest{
-			Database:   c.database,
-			Statements: ddlStatements,
-		})
-		if err != nil {
-			return nil, err
-		}
-		if err := op.Wait(ctx); err != nil {
-			return nil, err
-		}
-	}
-	return driver.ResultNoRows, nil
-}
-
-func (c *conn) execBatchDML(ctx context.Context, statements []spanner.Statement, options ExecOptions) (driver.Result, error) {
-	if len(statements) == 0 {
-		return &result{}, nil
-	}
-
-	var affected []int64
-	var err error
-	if c.inTransaction() {
-		tx, ok := c.tx.(*readWriteTransaction)
-		if !ok {
-			return nil, status.Errorf(codes.FailedPrecondition, "connection is in a transaction that is not a read/write transaction")
-		}
-		affected, err = tx.rwTx.BatchUpdateWithOptions(ctx, statements, options.QueryOptions)
-	} else {
-		_, err = c.client.ReadWriteTransactionWithOptions(ctx, func(ctx context.Context, transaction *spanner.ReadWriteTransaction) error {
-			affected, err = transaction.BatchUpdateWithOptions(ctx, statements, options.QueryOptions)
-			return err
-		}, options.TransactionOptions)
-	}
-	return &result{rowsAffected: sum(affected)}, err
-}
-
-func sum(affected []int64) int64 {
-	sum := int64(0)
-	for _, c := range affected {
-		sum += c
-	}
-	return sum
-}
-
-func (c *conn) Apply(ctx context.Context, ms []*spanner.Mutation, opts ...spanner.ApplyOption) (commitTimestamp time.Time, err error) {
-	if c.inTransaction() {
-		return time.Time{}, spanner.ToSpannerError(
-			status.Error(
-				codes.FailedPrecondition,
-				"Apply may not be called while the connection is in a transaction. Use BufferWrite to write mutations in a transaction."))
-	}
-	return c.client.Apply(ctx, ms, opts...)
-}
-
-func (c *conn) BufferWrite(ms []*spanner.Mutation) error {
-	if !c.inTransaction() {
-		return spanner.ToSpannerError(
-			status.Error(
-				codes.FailedPrecondition,
-				"BufferWrite may not be called while the connection is not in a transaction. Use Apply to write mutations outside a transaction."))
-	}
-	return c.tx.BufferWrite(ms)
-}
-
-// Ping implements the driver.Pinger interface.
-// returns ErrBadConn if the connection is no longer valid.
-func (c *conn) Ping(ctx context.Context) error {
-	if c.closed {
-		return driver.ErrBadConn
-	}
-	rows, err := c.QueryContext(ctx, "SELECT 1", []driver.NamedValue{})
-	if err != nil {
-		return driver.ErrBadConn
-	}
-	defer rows.Close()
-	values := make([]driver.Value, 1)
-	if err := rows.Next(values); err != nil {
-		return driver.ErrBadConn
-	}
-	if values[0] != int64(1) {
-		return driver.ErrBadConn
-	}
-	return nil
-}
-
-// ResetSession implements the driver.SessionResetter interface.
-// returns ErrBadConn if the connection is no longer valid.
-func (c *conn) ResetSession(_ context.Context) error {
-	if c.closed {
-		return driver.ErrBadConn
-	}
-	if c.inTransaction() {
-		if err := c.tx.Rollback(); err != nil {
-			return driver.ErrBadConn
-		}
-	}
-	c.commitTs = nil
-	c.batch = nil
-	c.retryAborts = true
-	c.autocommitDMLMode = Transactional
-	c.readOnlyStaleness = spanner.TimestampBound{}
-	return nil
-}
-
-// IsValid implements the driver.Validator interface.
-func (c *conn) IsValid() bool {
-	return !c.closed
-}
-
-// checkIsValidType returns true for types that do not need extra conversion
-// for spanner.
-func checkIsValidType(v driver.Value) bool {
-	switch v.(type) {
-	default:
-		return false
-	case nil:
-	case sql.NullInt64:
-	case sql.NullTime:
-	case sql.NullString:
-	case sql.NullFloat64:
-	case sql.NullBool:
-	case sql.NullInt32:
-	case string:
-	case spanner.NullString:
-	case []string:
-	case []spanner.NullString:
-	case *string:
-	case []*string:
-	case []byte:
-	case [][]byte:
-	case uint:
-	case []uint:
-	case *uint:
-	case []*uint:
-	case *[]uint:
-	case int:
-	case []int:
-	case *int:
-	case []*int:
-	case *[]int:
-	case int64:
-	case []int64:
-	case spanner.NullInt64:
-	case []spanner.NullInt64:
-	case *int64:
-	case []*int64:
-	case uint64:
-	case []uint64:
-	case *uint64:
-	case []*uint64:
-	case bool:
-	case []bool:
-	case spanner.NullBool:
-	case []spanner.NullBool:
-	case *bool:
-	case []*bool:
-	case float32:
-	case []float32:
-	case spanner.NullFloat32:
-	case []spanner.NullFloat32:
-	case *float32:
-	case []*float32:
-	case float64:
-	case []float64:
-	case spanner.NullFloat64:
-	case []spanner.NullFloat64:
-	case *float64:
-	case []*float64:
-	case big.Rat:
-	case []big.Rat:
-	case spanner.NullNumeric:
-	case []spanner.NullNumeric:
-	case *big.Rat:
-	case []*big.Rat:
-	case time.Time:
-	case []time.Time:
-	case spanner.NullTime:
-	case []spanner.NullTime:
-	case *time.Time:
-	case []*time.Time:
-	case civil.Date:
-	case []civil.Date:
-	case spanner.NullDate:
-	case []spanner.NullDate:
-	case *civil.Date:
-	case []*civil.Date:
-	case spanner.NullJSON:
-	case []spanner.NullJSON:
-	case spanner.GenericColumnValue:
-	}
-	return true
-}
-
-func (c *conn) CheckNamedValue(value *driver.NamedValue) error {
-	if value == nil {
-		return nil
-	}
-
-	if execOptions, ok := value.Value.(ExecOptions); ok {
-		c.execOptions = execOptions
-		return driver.ErrRemoveArgument
-	}
-
-	if checkIsValidType(value.Value) {
-		return nil
-	}
-
-	// Convert the value using the default sql driver. This uses driver.Valuer,
-	// if implemented, and falls back to reflection. If the converted value is
-	// a supported spanner type, use it. Otherwise, ignore any errors and
-	// continue checking other supported spanner specific types.
-	if v, err := driver.DefaultParameterConverter.ConvertValue(value.Value); err == nil {
-		if checkIsValidType(v) {
-			value.Value = v
-			return nil
-		}
-	}
-
-	// google-cloud-go/spanner knows how to deal with these
-	if isStructOrArrayOfStructValue(value.Value) || isAnArrayOfProtoColumn(value.Value) {
-		return nil
-	}
-
-	return spanner.ToSpannerError(status.Errorf(codes.InvalidArgument, "unsupported value type: %T", value.Value))
-}
-
-func (c *conn) Prepare(query string) (driver.Stmt, error) {
-	return c.PrepareContext(context.Background(), query)
-}
-
-func (c *conn) PrepareContext(ctx context.Context, query string) (driver.Stmt, error) {
-	execOptions := c.options()
-	parsedSQL, args, err := parseParameters(query)
-	if err != nil {
-		return nil, err
-	}
-	return &stmt{conn: c, query: parsedSQL, numArgs: len(args), execOptions: execOptions}, nil
-}
-
-func (c *conn) QueryContext(ctx context.Context, query string, args []driver.NamedValue) (driver.Rows, error) {
-	// Execute client side statement if it is one.
-	clientStmt, err := parseClientSideStatement(c, query)
-	if err != nil {
-		return nil, err
-	}
-	if clientStmt != nil {
-		return clientStmt.QueryContext(ctx, args)
-	}
-
-	execOptions := c.options()
-	return c.queryContext(ctx, query, execOptions, args)
-}
-
-func (c *conn) queryContext(ctx context.Context, query string, execOptions ExecOptions, args []driver.NamedValue) (driver.Rows, error) {
-	// Clear the commit timestamp of this connection before we execute the query.
-	c.commitTs = nil
-
-	stmt, err := prepareSpannerStmt(query, args)
-	if err != nil {
-		return nil, err
-	}
-	var iter rowIterator
-	if c.tx == nil {
-		statementType := detectStatementType(query)
-		if statementType == statementTypeDml {
-			// Use a read/write transaction to execute the statement.
-			var commitTs time.Time
-			iter, commitTs, err = c.execSingleQueryTransactional(ctx, c.client, stmt, execOptions)
-			if err != nil {
-				return nil, err
-			}
-			c.commitTs = &commitTs
-		} else {
-			// The statement was either detected as being a query, or potentially not recognized at all.
-			// In that case, just default to using a single-use read-only transaction and let Spanner
-			// return an error if the statement is not suited for that type of transaction.
-			iter = &readOnlyRowIterator{c.execSingleQuery(ctx, c.client, stmt, c.readOnlyStaleness, execOptions)}
-		}
-	} else {
-		iter = c.tx.Query(ctx, stmt, execOptions.QueryOptions)
-	}
-	return &rows{it: iter, decodeOption: execOptions.DecodeOption}, nil
-}
-
-func (c *conn) ExecContext(ctx context.Context, query string, args []driver.NamedValue) (driver.Result, error) {
-	// Execute client side statement if it is one.
-	stmt, err := parseClientSideStatement(c, query)
-	if err != nil {
-		return nil, err
-	}
-	if stmt != nil {
-		return stmt.ExecContext(ctx, args)
-	}
-	execOptions := c.options()
-	return c.execContext(ctx, query, execOptions, args)
-}
-
-func (c *conn) execContext(ctx context.Context, query string, execOptions ExecOptions, args []driver.NamedValue) (driver.Result, error) {
-	// Clear the commit timestamp of this connection before we execute the statement.
-	c.commitTs = nil
-
-	statementType := detectStatementType(query)
-	// Use admin API if DDL statement is provided.
-	if statementType == statementTypeDdl {
-		// Spanner does not support DDL in transactions, and although it is technically possible to execute DDL
-		// statements while a transaction is active, we return an error to avoid any confusion whether the DDL
-		// statement is executed as part of the active transaction or not.
-		if c.inTransaction() {
-			return nil, spanner.ToSpannerError(status.Errorf(codes.FailedPrecondition, "cannot execute DDL as part of a transaction"))
-		}
-		return c.execDDL(ctx, spanner.NewStatement(query))
-	}
-
-	ss, err := prepareSpannerStmt(query, args)
-	if err != nil {
-		return nil, err
-	}
-
-	var rowsAffected int64
-	var commitTs time.Time
-	if c.tx == nil {
-		if c.InDMLBatch() {
-			c.batch.statements = append(c.batch.statements, ss)
-		} else {
-			if c.autocommitDMLMode == Transactional {
-				rowsAffected, commitTs, err = c.execSingleDMLTransactional(ctx, c.client, ss, execOptions)
-				if err == nil {
-					c.commitTs = &commitTs
-				}
-			} else if c.autocommitDMLMode == PartitionedNonAtomic {
-				rowsAffected, err = c.execSingleDMLPartitioned(ctx, c.client, ss, execOptions)
-			} else {
-				return nil, status.Errorf(codes.FailedPrecondition, "connection in invalid state for DML statements: %s", c.autocommitDMLMode.String())
-			}
-		}
-	} else {
-		rowsAffected, err = c.tx.ExecContext(ctx, ss, execOptions.QueryOptions)
-	}
-	if err != nil {
-		return nil, err
-	}
-	return &result{rowsAffected: rowsAffected}, nil
-}
-
-// options returns and resets the ExecOptions for the next statement.
-func (c *conn) options() ExecOptions {
-	defer func() { c.execOptions = ExecOptions{} }()
-	return c.execOptions
-}
-
-func (c *conn) Close() error {
-	return c.connector.decreaseConnCount()
-}
-
-func noTransaction() error {
-	return status.Errorf(codes.FailedPrecondition, "connection does not have a transaction")
-}
-
-func (c *conn) resetTransactionForRetry(ctx context.Context, errDuringCommit bool) error {
-	if errDuringCommit {
-		if c.prevTx == nil {
-			return noTransaction()
-		}
-		c.tx = c.prevTx
-		c.resetForRetry = true
-	} else if c.tx == nil {
-		return noTransaction()
-	}
-	return c.tx.resetForRetry(ctx)
-}
-
-func (c *conn) withTransactionOptions(options spanner.TransactionOptions) {
-	c.execOptions.TransactionOptions = options
-}
-
-type spannerIsolationLevel sql.IsolationLevel
-
-const (
-	levelNone spannerIsolationLevel = iota
-	levelDisableRetryAborts
-)
-
-// WithDisableRetryAborts returns a specific Spanner isolation level that contains
-// both the given standard isolation level and a custom Spanner isolation level that
-// disables internal retries for aborted transactions for a single transaction.
-func WithDisableRetryAborts(level sql.IsolationLevel) sql.IsolationLevel {
-	return sql.IsolationLevel(levelDisableRetryAborts)<<8 + level
-}
-
-func (c *conn) Begin() (driver.Tx, error) {
-	return c.BeginTx(context.Background(), driver.TxOptions{})
-}
-
-func (c *conn) BeginTx(ctx context.Context, opts driver.TxOptions) (driver.Tx, error) {
-	if c.resetForRetry {
-		c.resetForRetry = false
-		return c.tx, nil
-	}
-	if c.inTransaction() {
-		return nil, spanner.ToSpannerError(status.Errorf(codes.FailedPrecondition, "already in a transaction"))
-	}
-	if c.inBatch() {
-		return nil, status.Error(codes.FailedPrecondition, "This connection has an active batch. Run or abort the batch before starting a new transaction.")
-	}
-
-	execOptions := c.options()
-	disableRetryAborts := false
-	sil := opts.Isolation >> 8
-	opts.Isolation = opts.Isolation - sil
-	if sil > 0 {
-		switch spannerIsolationLevel(sil) {
-		case levelDisableRetryAborts:
-			disableRetryAborts = true
-		}
-	}
-
-	if opts.ReadOnly {
-		logger := c.logger.With("tx", "ro")
-		ro := c.client.ReadOnlyTransaction().WithTimestampBound(c.readOnlyStaleness)
-		c.tx = &readOnlyTransaction{
-			roTx:   ro,
-			logger: logger,
-			close: func() {
-				c.tx = nil
-			},
-		}
-		return c.tx, nil
-	}
-
-	tx, err := spanner.NewReadWriteStmtBasedTransactionWithOptions(ctx, c.client, execOptions.TransactionOptions)
-	if err != nil {
-		return nil, err
-	}
-	logger := c.logger.With("tx", "rw")
-	c.tx = &readWriteTransaction{
-		ctx:    ctx,
-		client: c.client,
-		logger: logger,
-		rwTx:   tx,
-		close: func(commitTs *time.Time, commitErr error) {
-			c.prevTx = c.tx
-			c.tx = nil
-			if commitErr == nil {
-				c.commitTs = commitTs
-			}
-		},
-		retryAborts: c.retryAborts && !disableRetryAborts,
-	}
-	c.commitTs = nil
-	return c.tx, nil
-}
-
-func (c *conn) inTransaction() bool {
-	return c.tx != nil
-}
-
-func (c *conn) inReadOnlyTransaction() bool {
-	if c.tx != nil {
-		_, ok := c.tx.(*readOnlyTransaction)
-		return ok
-	}
-	return false
-}
-
-func (c *conn) inReadWriteTransaction() bool {
-	if c.tx != nil {
-		_, ok := c.tx.(*readWriteTransaction)
-		return ok
-	}
-	return false
-}
-
-func queryInSingleUse(ctx context.Context, c *spanner.Client, statement spanner.Statement, tb spanner.TimestampBound, options ExecOptions) *spanner.RowIterator {
-	return c.Single().WithTimestampBound(tb).QueryWithOptions(ctx, statement, options.QueryOptions)
-}
-
-type wrappedRowIterator struct {
-	*spanner.RowIterator
-
-	noRows   bool
-	firstRow *spanner.Row
-}
-
-func (ri *wrappedRowIterator) Next() (*spanner.Row, error) {
-	if ri.noRows {
-		return nil, iterator.Done
-	}
-	if ri.firstRow != nil {
-		defer func() { ri.firstRow = nil }()
-		return ri.firstRow, nil
-	}
-	return ri.RowIterator.Next()
-}
-
-func (ri *wrappedRowIterator) Stop() {
-	ri.RowIterator.Stop()
-}
-
-func (ri *wrappedRowIterator) Metadata() *spannerpb.ResultSetMetadata {
-	return ri.RowIterator.Metadata
-}
-
 func queryInNewRWTransaction(ctx context.Context, c *spanner.Client, statement spanner.Statement, options ExecOptions) (rowIterator, time.Time, error) {
 	var result *wrappedRowIterator
 	fn := func(ctx context.Context, tx *spanner.ReadWriteTransaction) error {
@@ -1592,32 +640,3 @@ func execAsPartitionedDML(ctx context.Context, c *spanner.Client, statement span
 	queryOptions.ExcludeTxnFromChangeStreams = options.TransactionOptions.ExcludeTxnFromChangeStreams
 	return c.PartitionedUpdateWithOptions(ctx, statement, queryOptions)
 }
-
-/* The following is the same implementation as in google-cloud-go/spanner */
-
-func isStructOrArrayOfStructValue(v interface{}) bool {
-	typ := reflect.TypeOf(v)
-	if typ.Kind() == reflect.Slice {
-		typ = typ.Elem()
-	}
-	if typ.Kind() == reflect.Ptr {
-		typ = typ.Elem()
-	}
-	return typ.Kind() == reflect.Struct
-}
-
-func isAnArrayOfProtoColumn(v interface{}) bool {
-	typ := reflect.TypeOf(v)
-	if typ.Kind() == reflect.Ptr {
-		typ = typ.Elem()
-	}
-	if typ.Kind() == reflect.Slice {
-		typ = typ.Elem()
-	}
-	return typ.Implements(protoMsgReflectType) || typ.Implements(protoEnumReflectType)
-}
-
-var (
-	protoMsgReflectType  = reflect.TypeOf((*proto.Message)(nil)).Elem()
-	protoEnumReflectType = reflect.TypeOf((*protoreflect.Enum)(nil)).Elem()
-)
